bullets formula correct. added colission. need to implement it better

+++ server/cmd/bullet.go
@@ -9,17 +9,17 @@ import (
 type Bullet struct {
 	ID         string    `json:"id"`
 	Position   *Position `json:"position"`
-	Angle      float32   `json:"angle"`
+	Angle      float64   `json:"angle"`
 	BulletType string    `json:"bullet_type"`
-	Speed      float32   `json:"speed"`
+	Speed      float64   `json:"speed"`
 }
 
-func NewBullet(position *Position, angle float32, bulletType string) *Bullet {
+func NewBullet(position *Position, angle float64, bulletType string) *Bullet {
 	return &Bullet{ID: uuid.New().String(), Position: position, Angle: angle, BulletType: bulletType, Speed: 300}
 }
 
-func (b *Bullet) update(dt float32) {
-	rad := float64(b.Angle)
-	b.Position.X += float32(math.Cos(rad)) * b.Speed * dt
-	b.Position.Y += float32(math.Sin(rad)) * b.Speed * dt
+func (b *Bullet) update(dt float64) {
+	rad := b.Angle
+	b.Position.X += math.Cos(rad) * b.Speed * dt
+	b.Position.Y += math.Sin(rad) * b.Speed * dt
 }
